tests fix
